Character Structures

IDS files use Unicode ideographic description characters to describe how components of characters are put together (they occupy the range from U+2FF0 to U+2FFB). In IDSFiles.jl, these relationships are represented as a character structures, similar to a parse tree.

IDSFiles.parse — Method

IDSFiles.parse(::AbstractString)

Parse a single IDS string that describes a character/component's decomposition. (This function does not do any input validation to verify that the string can actually represent a decomposition; in this case, the method will fail.)

This is a relatively low-level method; you typically do not need to call this directly.

Examples

We can take the IDS string itself and parse it as a nested structure.

julia> parse("⿱此二")
CharStructure{TopBottomStructure}(TopBottomStructure(Component('此'), Component('二')))

Structure Hierarchy

Missing docstring.

Missing docstring for AbstractCharStructure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NestedStructure. Check Documenter's build log for details.

IDSFiles.Component — Type

Leaf node of a structure tree; this contains a single component.

Basic Relationships

Missing docstring.

Missing docstring for LeftRightStructure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TopBottomStructure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for VerticalThirdsStructure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HorizontalThirdsStructure. Check Documenter's build log for details.

Nested Relationships

Missing docstring.

Missing docstring for CenterNestedStructure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BottomNestedStructure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TopNestedStructure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RightNestedStructure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BottomRightNestedStructure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BottomLeftNestedStructure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TopRightNestedStructure. Check Documenter's build log for details.

Overlap Relationship

Missing docstring.

Missing docstring for OverlapStructure. Check Documenter's build log for details.